Adding conditional symbolic link creation

+++ scripts/after_install.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 set -ex
 
-sudo ln -s /opt/apps/jdb/nginx.conf /etc/nginx/sites-enabled/jdb
+if [ ! -f /etc/nginx/sites-enabled/jdb ]; then
+  sudo ln -s /opt/apps/jdb/nginx.conf /etc/nginx/sites-enabled/jdb
+fi
 # sudo ln -s /opt/apps/jdb/uwsgi.ini /etc/uwsgi/apps-enabled/jdb.ini
 
 sudo service nginx restart
